+++ src/main/java/com/teamscale/upload/autodetect_revision/ProcessUtils.java
@@ -1,18 +1,10 @@
 package com.teamscale.upload.autodetect_revision;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 
-import org.apache.commons.exec.CommandLine;
-import org.apache.commons.exec.DefaultExecutor;
-import org.apache.commons.exec.ExecuteStreamHandler;
-import org.apache.commons.exec.PumpStreamHandler;
+import com.teamscale.upload.utils.FileSystemUtils;
 
 /**
  * Utility methods for executing processes on the command line.
@@ -27,107 +19,140 @@ public class ProcessUtils {
 
 	private static final int EXIT_CODE_SUCCESS = 0;
 
-	/** Run the command with the given arguments. */
-	public static ProcessResult run(String command, String... arguments) {
-		return runWithStdin(command, null, arguments);
+	/**
+	 * Starts a {@link Process} for the command and returns the
+	 * {@link ProcessResult}.
+	 */
+	public static ProcessResult run(String... command) {
+		return runWithStdIn(null, command);
 	}
 
 	/**
-	 * Run the command with the given arguments. This method should not be called
-	 * directly in production code (only with parameter stdinFile=null).
-	 *
-	 * To allow simulating user input in test, this method takes a file which can be
-	 * used to pipe input to stdin of the command. The parameter stdinFile may be
-	 * null to indicate that no stdin should be used.
+	 * Starts a {@link Process} for the command and returns the
+	 * {@link ProcessResult}. Additionally, takes a file which can be used to pipe
+	 * input to stdin of the command. The parameter stdinFile may be null to
+	 * indicate that no stdin should be used.
 	 */
-	public static ProcessResult runWithStdin(String command, String stdinFile, String... arguments) {
+	public static ProcessResult runWithStdIn(File stdInFile, String... command) {
+		try {
+			ProcessBuilder processBuilder = new ProcessBuilder(command);
+			if (stdInFile != null) {
+				processBuilder.redirectInput(stdInFile);
+			}
+			Process process = processBuilder.start();
 
-		CommandLine commandLine = new CommandLine(command);
-		commandLine.addArguments(arguments, false);
+			/*
+			 * Both input streams need to be drained in separate threads since reading
+			 * either stream can be a blocking operation if the other stream has reached the
+			 * platform dependent buffer limit for standard output.
+			 *
+			 * See https://stackoverflow.com/a/7562321.
+			 */
+			ProcessOutputReader inputStreamReader = new ProcessOutputReader(process.getInputStream());
+			ProcessOutputReader errorStreamReader = new ProcessOutputReader(process.getErrorStream());
+			Thread inputStreamReaderThread = new Thread(inputStreamReader);
+			Thread errorStreamReaderThread = new Thread(errorStreamReader);
+			inputStreamReaderThread.start();
+			errorStreamReaderThread.start();
 
-		ByteArrayInputStream input = null;
-		if (stdinFile != null) {
+			int exitCode = process.waitFor();
+
+			/*
+			 * Ensure that both threads have finished execution if the process terminates
+			 * earlier than the threads.
+			 */
+			inputStreamReaderThread.join();
+			errorStreamReaderThread.join();
+			inputStreamReader.rethrowCaughtException();
+			errorStreamReader.rethrowCaughtException();
+
+			return new ProcessResult(exitCode, inputStreamReader.result, errorStreamReader.result, null);
+		} catch (IOException | InterruptedException e) {
+			return new ProcessResult(-1, "", e.getMessage(), e);
+		}
+	}
+
+	/**
+	 * Runnable for reading the input stream asynchronously from a separate thread
+	 * to prevent a deadlock due to blocking read operations.
+	 */
+	private static class ProcessOutputReader implements Runnable {
+
+		private final InputStream inputStream;
+
+		private IOException exception;
+
+		private String result;
+
+		private ProcessOutputReader(InputStream inputStream) {
+			this.inputStream = inputStream;
+		}
+
+		@Override
+		public void run() {
 			try {
-				// We cannot directly pipe the file output to stdin, but we can use a byte array
-				// input stream
-				// https://stackoverflow.com/questions/4695664/how-to-pipe-a-string-argument-to-an-executable-launched-with-apache-commons-exec
-				String accessKey = Files.readString(Paths.get(stdinFile));
-				input = new ByteArrayInputStream(accessKey.getBytes(StandardCharsets.ISO_8859_1));
+				result = FileSystemUtils.getInputAsString(inputStream);
 			} catch (IOException e) {
-				System.err.println("Could not read access key from file `" + stdinFile + ".");
-				return new ProcessResult(-1, "", e);
+				exception = e;
 			}
 		}
 
-		DefaultExecutor executor = new DefaultExecutor();
-		CaptureStreamHandler handler = new CaptureStreamHandler(input);
-		executor.setStreamHandler(handler);
-		executor.setExitValues(null); // don't throw in case of non-zero exit codes
-
-		try {
-			int exitCode = executor.execute(commandLine);
-			return new ProcessResult(exitCode, handler.getStdOutAndStdErr(), null);
-		} catch (IOException e) {
-			System.err.println("Tried to run `" + command + " " + String.join(" ", arguments)
-					+ "` which failed with an exception");
-			e.printStackTrace();
-			return new ProcessResult(-1, "", e);
+		private void rethrowCaughtException() throws IOException {
+			if (exception != null) {
+				throw exception;
+			}
 		}
 	}
 
-	private static class CaptureStreamHandler implements ExecuteStreamHandler {
-
-		private final ByteArrayOutputStream stdoutAndStderrStream = new ByteArrayOutputStream();
-		private final PumpStreamHandler wrappedHandler;
-
-		public CaptureStreamHandler(ByteArrayInputStream input) {
-			// Currently, we do not need to differentiate between stdout and stderr
-			wrappedHandler = new PumpStreamHandler(stdoutAndStderrStream, stdoutAndStderrStream, input);
-		}
-
-		public String getStdOutAndStdErr() {
-			// we want to use the platform default charset here
-			// as I'm guessing it's used for the process output
-			return stdoutAndStderrStream.toString();
-		}
-
-		@Override
-		public void setProcessInputStream(OutputStream os) {
-			wrappedHandler.setProcessInputStream(os);
-		}
-
-		@Override
-		public void setProcessErrorStream(InputStream is) {
-			wrappedHandler.setProcessErrorStream(is);
-		}
-
-		@Override
-		public void setProcessOutputStream(InputStream is) {
-			wrappedHandler.setProcessOutputStream(is);
-		}
-
-		@Override
-		public void start() {
-			wrappedHandler.start();
-		}
-
-		@Override
-		public void stop() throws IOException {
-			wrappedHandler.stop();
-		}
-	}
-
+	/**
+	 * The result of a process execution.
+	 */
 	public static class ProcessResult {
-		public final int exitCode;
-		public final String stdoutAndStdErr;
-		public final IOException exception;
 
-		private ProcessResult(int exitCode, String stdoutAndStdErr, IOException exception) {
+		/**
+		 * The exit code of the process.
+		 */
+		public final int exitCode;
+
+		/**
+		 * The stdout output of the process.
+		 */
+		public final String output;
+
+		/**
+		 * The stderr output of the process.
+		 */
+		public final String errorOutput;
+
+		/**
+		 * A potential Java exception that occurred while executing the process.
+		 */
+		public final Exception exception;
+
+		private ProcessResult(int exitCode, String output, String errorOutput, Exception exception) {
 			this.exitCode = exitCode;
-			this.stdoutAndStdErr = stdoutAndStdErr;
+			this.output = output;
+			this.errorOutput = errorOutput;
 			this.exception = exception;
 		}
 
+		/**
+		 * Returns the {@link #output} followed by the {@link #errorOutput}.
+		 */
+		public String getOutputAndErrorOutput() {
+			String outputAndErrorOutput = "";
+			if (output != null) {
+				outputAndErrorOutput += output;
+			}
+			if (errorOutput != null) {
+				outputAndErrorOutput += errorOutput;
+			}
+			return outputAndErrorOutput;
+		}
+
+		/**
+		 * Returns true if the processed exited successfully.
+		 */
 		public boolean wasSuccessful() {
 			return exception == null && exitCode == EXIT_CODE_SUCCESS;
 		}