+++ src/main/java/com/teamscale/upload/xcode/XCResultConverter.java
@@ -8,7 +8,6 @@ import java.nio.file.Files;
 import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Queue;
@@ -20,15 +19,9 @@ import java.util.concurrent.TimeUnit;
 
 import com.google.gson.Gson;
 import com.teamscale.upload.autodetect_revision.ProcessUtils;
-import com.teamscale.upload.report.testwise_coverage.TestInfo;
-import com.teamscale.upload.report.testwise_coverage.TestwiseCoverageReport;
+import com.teamscale.upload.autodetect_revision.ProcessUtils.ProcessResult;
 import com.teamscale.upload.report.xcode.ActionRecord;
-import com.teamscale.upload.report.xcode.ActionTest;
-import com.teamscale.upload.report.xcode.ActionTestPlanRunSummaries;
-import com.teamscale.upload.report.xcode.ActionTestPlanRunSummary;
-import com.teamscale.upload.report.xcode.ActionTestableSummary;
 import com.teamscale.upload.report.xcode.ActionsInvocationRecord;
-import com.teamscale.upload.report.xcode.XCResultObjectIdReference;
 import com.teamscale.upload.utils.FileSystemUtils;
 import com.teamscale.upload.utils.LogUtils;
 
@@ -44,11 +37,6 @@ public class XCResultConverter {
 	public static final String XCODE_REPORT_FORMAT = "XCODE";
 
 	/**
-	 * The enum name of the Testwise Coverage report format.
-	 */
-	private static final String TESTWISE_COVERAGE_REPORT_FORMAT = "TESTWISE_COVERAGE";
-
-	/**
 	 * The number of conversion threads to run in parallel for faster conversion. By
 	 * default, we use the number of available processors to distribute work since
 	 * this setting was most performant when testing locally.
@@ -60,11 +48,6 @@ public class XCResultConverter {
 	 * File extension used for converted XCResult bundles.
 	 */
 	private static final String XCCOV_REPORT_FILE_EXTENSION = ".xccov";
-
-	/**
-	 * File extension used for converted XCResult bundles.
-	 */
-	private static final String TESTWISE_COVERAGE_REPORT_FILE_EXTENSION = ".testwisecoverage.json";
 
 	private final File workingDirectory;
 
@@ -105,8 +88,6 @@ public class XCResultConverter {
 	 */
 	private static void writeConversionResults(Queue<Future<ConversionResult>> conversionResults, File convertedReport)
 			throws InterruptedException, ExecutionException, IOException {
-		FileSystemUtils.ensureEmptyFile(convertedReport);
-
 		while (!conversionResults.isEmpty()) {
 			ConversionResult conversionResult = conversionResults.remove().get();
 			if (conversionResult == null) {
@@ -125,10 +106,15 @@ public class XCResultConverter {
 	 * Returns a sorted list of source files contained in the XCResult bundle
 	 * directory.
 	 */
-	private static List<String> getSourceFiles(File reportDirectory) throws IOException, InterruptedException {
-		String output = ProcessUtils.run("xcrun", "xccov", "view", "--archive", "--file-list",
-				reportDirectory.getAbsolutePath()).stdoutAndStdErr;
-		return output.lines().sorted().collect(toList());
+	private static List<String> getSourceFiles(File reportDirectory) throws ConversionException {
+		ProcessResult result = ProcessUtils.run("xcrun", "xccov", "view", "--archive", "--file-list",
+				reportDirectory.getAbsolutePath());
+		if (!result.wasSuccessful()) {
+			throw new ConversionException(
+					"Error while obtaining file list from XCResult archive " + reportDirectory.getAbsolutePath(),
+					result.exception);
+		}
+		return result.output.lines().sorted().collect(toList());
 	}
 
 	private static void validateCommandLineTools() throws IOException, InterruptedException, ConversionException {
@@ -141,8 +127,7 @@ public class XCResultConverter {
 
 	/**
 	 * Converts the report and writes the conversion result to a file with the same
-	 * path and the {@link #XCCOV_REPORT_FILE_EXTENSION} or
-	 * {@link #TESTWISE_COVERAGE_REPORT_FILE_EXTENSION} as an added file extension.
+	 * path and the {@link #XCCOV_REPORT_FILE_EXTENSION} as an added file extension.
 	 */
 	public List<ConvertedReport> convert(File report) throws ConversionException {
 		try {
@@ -155,22 +140,19 @@ public class XCResultConverter {
 				convertedReports.add(extractCoverageData(report, reportDirectory));
 			} else {
 				ActionsInvocationRecord actionsInvocationRecord = getActionsInvocationRecord(reportDirectory);
-				convertedReports.add(extractTestResults(report, reportDirectory, actionsInvocationRecord));
 
 				if (actionsInvocationRecord.hasCoverageData()) {
 					for (File convertToXccovArchive : convertToXccovArchives(reportDirectory,
 							actionsInvocationRecord)) {
 						convertedReports.add(extractCoverageData(report, convertToXccovArchive));
 					}
+				} else {
+					LogUtils.warn("XCResult bundle doesn't contain any coverage data.");
 				}
 			}
 
 			return convertedReports;
-		} catch (IOException e) {
-			LogUtils.warn(
-					String.format("Error while converting report %s: %s", report.getAbsolutePath(), e.getMessage()));
-			return Collections.emptyList();
-		} catch (InterruptedException | ExecutionException e) {
+		} catch (IOException | InterruptedException | ExecutionException e) {
 			throw new ConversionException(
 					String.format("Error while converting report %s: %s", report.getAbsolutePath(), e.getMessage()), e);
 		}
@@ -221,72 +203,44 @@ public class XCResultConverter {
 	}
 
 	private ActionsInvocationRecord getActionsInvocationRecord(File reportDirectory)
-			throws IOException, InterruptedException {
-		String actionsInvocationRecordJson = ProcessUtils.run("xcrun", "xcresulttool", "get", "--path",
-				reportDirectory.getAbsolutePath(), "--format", "json").stdoutAndStdErr;
+			throws InterruptedException, ConversionException {
+		ProcessResult result = ProcessUtils.run("xcrun", "xcresulttool", "get", "--path",
+				reportDirectory.getAbsolutePath(), "--format", "json");
+		if (!result.wasSuccessful()) {
+			throw new ConversionException("Error while obtaining ActionInvocationsRecord from XCResult archive "
+					+ reportDirectory.getAbsolutePath(), result.exception);
+		}
+		String actionsInvocationRecordJson = result.output;
 		return new Gson().fromJson(actionsInvocationRecordJson, ActionsInvocationRecord.class);
 	}
 
-	private ConvertedReport extractTestResults(File report, File reportDirectory,
-			ActionsInvocationRecord actionsInvocationRecord) throws IOException, InterruptedException {
-		List<TestInfo> tests = new ArrayList<>();
-
-		for (ActionRecord action : actionsInvocationRecord.actions) {
-			XCResultObjectIdReference testsRef = action.actionResult.testsRef;
-			if (testsRef == null) {
-				continue;
-			}
-
-			String json = ProcessUtils.run("xcrun", "xcresulttool", "get", "--path", reportDirectory.getAbsolutePath(),
-					"--format", "json", "--id", testsRef.id).stdoutAndStdErr;
-			ActionTestPlanRunSummaries actionTestPlanRunSummaries = new Gson().fromJson(json,
-					ActionTestPlanRunSummaries.class);
-
-			for (ActionTestPlanRunSummary summary : actionTestPlanRunSummaries.summaries) {
-				for (ActionTestableSummary testableSummary : summary.testableSummaries) {
-					for (ActionTest test : testableSummary.tests) {
-						extractTests(test, tests);
-					}
-				}
-			}
-		}
-
-		File testwiseCoverageReportFile = new File(report.getAbsolutePath() + TESTWISE_COVERAGE_REPORT_FILE_EXTENSION);
-
-		tests.sort(Comparator.comparing(testInfo -> testInfo.uniformPath));
-		TestwiseCoverageReport testwiseCoverageReport = new TestwiseCoverageReport(tests);
-
-		FileSystemUtils.ensureEmptyFile(testwiseCoverageReportFile);
-		Files.writeString(testwiseCoverageReportFile.toPath(), new Gson().toJson(testwiseCoverageReport));
-
-		return new ConvertedReport(TESTWISE_COVERAGE_REPORT_FORMAT, testwiseCoverageReportFile);
-	}
-
-	private void extractTests(ActionTest actionTest, List<TestInfo> tests) {
-		if (actionTest.subTests == null) {
-			TestInfo.create(actionTest).ifPresent(tests::add);
-			return;
-		}
-
-		for (ActionTest subTest : actionTest.subTests) {
-			extractTests(subTest, tests);
-		}
-	}
-
 	private ConvertedReport extractCoverageData(File report, File reportDirectory)
-			throws IOException, InterruptedException, ExecutionException {
+			throws IOException, InterruptedException, ExecutionException, ConversionException {
 		List<String> sourceFiles = getSourceFiles(reportDirectory);
-
-		LogUtils.info(String.format("Extracting coverage for %d source files using %d threads from XCResult bundle %s.",
-				sourceFiles.size(), CONVERSION_THREAD_COUNT, report.getAbsolutePath()));
-
+		File convertedCoverageReport = new File(report.getAbsolutePath() + XCCOV_REPORT_FILE_EXTENSION);
 		long startTime = System.currentTimeMillis();
 
-		Queue<Future<ConversionResult>> conversionResults = submitConversionTasks(reportDirectory, sourceFiles);
-		File convertedCoverageReport = new File(report.getAbsolutePath() + XCCOV_REPORT_FILE_EXTENSION);
+		LogUtils.info(String.format("Converting XCResult bundle %s containing %d source files.",
+				report.getAbsolutePath(), sourceFiles.size()));
+		ProcessResult result = ProcessUtils.run("xcrun", "xccov", "view", "--archive",
+				reportDirectory.getAbsolutePath());
+		FileSystemUtils.ensureEmptyFile(convertedCoverageReport);
 
-		writeConversionResults(conversionResults, convertedCoverageReport);
-		waitForRunningProcessesToFinish();
+		/*
+		 * With XCode 13.3 and newer the coverage of all source files can be exported
+		 * with the command above which offers the best performance. We try this first
+		 * and if it doesn't work we fall back to the slower legacy mechanism that
+		 * iterates over each source file.
+		 */
+		if (result.wasSuccessful()) {
+			Files.writeString(convertedCoverageReport.toPath(), result.output, StandardOpenOption.WRITE);
+		} else {
+			LogUtils.info(String.format("Using legacy conversion with %d threads.", CONVERSION_THREAD_COUNT));
+			Queue<Future<ConversionResult>> conversionResults = submitConversionTasks(reportDirectory, sourceFiles);
+
+			writeConversionResults(conversionResults, convertedCoverageReport);
+			waitForRunningProcessesToFinish();
+		}
 
 		LogUtils.info(String.format("Coverage extraction finished after %d seconds.",
 				(System.currentTimeMillis() - startTime) / 1000));
@@ -307,19 +261,25 @@ public class XCResultConverter {
 		return conversionResults;
 	}
 
-	private void waitForRunningProcessesToFinish() throws InterruptedException {
-		if (executorService != null && !executorService.isTerminated()) {
+	private void waitForRunningProcessesToFinish() throws ConversionException {
+		if (executorService == null || executorService.isTerminated()) {
+			return;
+		}
+		try {
 			if (!executorService.awaitTermination(60, TimeUnit.SECONDS)) {
 				LogUtils.warn("Processes took too long to terminate. Forcing shutdown.");
 				executorService.shutdownNow();
 			}
+		} catch (InterruptedException e) {
+			throw new ConversionException(
+					"Error occurred while waiting for conversion processes to finish: " + e.getMessage(), e);
 		}
 	}
 
 	/**
 	 * Stops the conversion processes and ensures that all processes have finished.
 	 */
-	public void stopProcesses() throws InterruptedException, IOException {
+	public void stopProcesses() throws ConversionException {
 		if (executorService != null) {
 			if (!executorService.isTerminated()) {
 				executorService.shutdownNow();
@@ -333,11 +293,11 @@ public class XCResultConverter {
 	 */
 	public static class ConversionException extends Exception {
 
-		private ConversionException(String message) {
+		public ConversionException(String message) {
 			super(message);
 		}
 
-		private ConversionException(String message, Exception e) {
+		public ConversionException(String message, Exception e) {
 			super(message, e);
 		}
 	}